==> Figure 1/Step5_FCG_subclustering.R <==
#install.packages
library(Seurat)
library(ggplot2)
library(DoubletFinder)
library(dplyr)
library(cowplot)
library(reshape2)
library(MAST)
setwd("/athena/ganlab/scratch/lif4001/FCG/integration")
MG <- readRDS("FCG_MG_subset.rds")
OL <- readRDS("FCG_OL_subset.rds")
#######################################################################
setwd("/athena/ganlab/scratch/lif4001/FCG/integration/MG_only")
DefaultAssay(MG) <- 'integrated'
MG <- ScaleData(MG, verbose = FALSE)
MG <- RunPCA(MG, features = VariableFeatures(object = MG), verbose = FALSE)
ElbowPlot(MG)
MG <- FindNeighbors(MG, dims = 1:15)
MG <- FindClusters(MG, resolution = 0.2)
MG <- RunUMAP(MG, dims = 1: 15)
# rename cluster
n <- dim(table(MG@active.ident))
current.cluster.ids <- c(0, seq(1:(n-1)))
new.cluster.ids <- current.cluster.ids + 1
MG@active.ident <- plyr::mapvalues(x = MG@active.ident, from = current.cluster.ids, to = new.cluster.ids)
MG@active.ident <- factor(MG@active.ident, levels=1:n)
saveRDS(MG, file = 'FCG_MG_reclusted_res0.2.rds')
pdf("FCG_MG_umap.pdf", width=3.3, height=2.7)
DimPlot(MG, reduction = 'umap', label = T)
dev.off()
pdf("FCG_MG_umap_Condition.pdf", width=10, height=4.5)
DimPlot(MG, reduction = "umap", split.by = "Genotype.Food", label = T, ncol = 4)
dev.off()
pdf("FCG_MG_umap_Sample.pdf", width=12, height=7)
DimPlot(MG, reduction = "umap", split.by = "Sample", label = T, ncol = 6)
dev.off()
DefaultAssay(MG) <- 'RNA'
FCG_MG_markers <- FindAllMarkers(MG, logfc.threshold = 0.1, test.use = "MAST",min.pct = 0.25, only.pos = T)
write.csv(FCG_MG_markers, "FCG_MG_markers.csv")
write.csv(table(MG$seurat_clusters, MG$Sample_Name), "MG_subcluster_cell_counts.csv")
#######################################################################
setwd("/athena/ganlab/scratch/lif4001/FCG/integration/OL_only")
DefaultAssay(OL) <- 'integrated'
OL <- ScaleData(OL, verbose = FALSE)
OL <- RunPCA(OL, features = VariableFeatures(object = OL), verbose = FALSE)
ElbowPlot(OL)
OL <- FindNeighbors(OL, dims = 1:15)
OL <- FindClusters(OL, resolution = 0.15)
OL <- RunUMAP(OL, dims = 1: 15)
# rename cluster
n <- dim(table(OL@active.ident))
current.cluster.ids <- c(0, seq(1:(n-1)))
new.cluster.ids <- current.cluster.ids + 1
OL@active.ident <- plyr::mapvalues(x = OL@active.ident, from = current.cluster.ids, to = new.cluster.ids)
OL@active.ident <- factor(OL@active.ident, levels=1:n)
saveRDS(OL, file = 'FCG_OL_reclusted_res0.15.rds')
pdf("FCG_OL_umap.pdf", width=3.3, height=2.7)
DimPlot(OL, reduction = 'umap', label = T)
dev.off()
pdf("FCG_OL_umap_Condition.pdf", width=10, height=4.5)
DimPlot(OL, reduction = "umap", split.by = "Genotype.Food", label = T, ncol = 4)
dev.off()
pdf("FCG_OL_umap_Sample.pdf", width=12, height=7)
DimPlot(OL, reduction = "umap", split.by = "Sample", label = T, ncol = 6)
dev.off()
DefaultAssay(OL) <- 'RNA'
FCG_OL_markers <- FindAllMarkers(OL, logfc.threshold = 0.1, test.use = "MAST",min.pct = 0.25, only.pos = T)
write.csv(FCG_OL_markers, "FCG_OL_markers.csv")
write.csv(table(OL$seurat_clusters, OL$Sample_Name), "OL_subcluster_cell_counts.csv")
